requirements.txt
torch
torchvision
tqdm
numpy
tifffile
Pillow
matplotlib
tensorboard
